Create README.md
manadatory execution of anyone who access proof of reserve not the owner no permission, locate 20 identification and sub ID of all unauthorized log in request video call confirmation of all request by batch on api and proof of reserve. Assylum cast and reposition from Protection fund from the neglighence of the exchange of any alert and cofirmation and acknowledgement.